Moved all the helper functions into helper.go
